Delete unused configuration files and scripts
