--- airflow/.env
@@ -1,3 +1,0 @@
-AIRFLOW_UID=50000
-
-
