--- airflow/scripts/git_scripts/.env
@@ -1,7 +1,0 @@
-GITHUB_TOKEN=token
-DATABASE_URL=postgresql+psycopg2://postgres:postgres@172.21.1.5:5432/git_db
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=postgres
-POSTGRES_DB=git_db
-PGADMIN_DEFAULT_EMAIL=admin@admin.com
-PGADMIN_DEFAULT_PASSWORD=admin
