+++ discord/.dockerignore
