+++ discord/requirements.txt
