--- elastic/logstash-config/logstash.yml
@@ -1,4 +1,0 @@
-pipeline.id: main
-path.config: "/usr/share/logstash/pipeline/logstash.conf"
-
-
