--- elastic/logstash-pipeline/logstash.conf
@@ -1,12 +1,0 @@
-input {
-  beats {
-    port => 5044
-  }
-}
-
-output {
-  elasticsearch {
-    hosts => ["elasticsearch:9200"]
-  }
-}
-
